fix React placement
